Added test that validates ruleset is loaded from csproj.
